Clases - Peticiones api - Se borro la clase "RestCall...". - Se crearon las clase "Api...", "ResponseApi..." y "RequestApi..." que reemplazaran a la clase "RestCall...".
