+++ App/Facades/Api/RequestApiFacade.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace App\Facades\Api;
+
+use Silver\Support\Facade;
+
+
+/**
+ * RequestApiFacade event provider
+ */
+class RequestApiFacade extends Facade
+{
+
+    protected static function getClass()
+    {
+        return 'App\Helpers\Api\RequestApiHelper';
+    }
+
+}
