--- App/Facades/Api/RestCallFacade.php
@@ -1,16 +1,0 @@
-<?php
-
-namespace App\Facades\Api;
-
-use Silver\Support\Facade;
-
-/**
- * RestCallFacade event provider
- */
-class RestCallFacade extends Facade {
-    
-    protected static function getClass() {
-        return 'App\Helpers\Api\RestCallHelper';
-    }
-    
-}
