+++ App/Helpers/Api/RequestApiHelper.php
@@ -1,41 +1,26 @@
 <?php
 /**
- * RestCallHelper.php
+ * RequestApiHelper.php
  */
 
 namespace App\Helpers\Api;
 
-use App\Facades\TokenFacade;
-use App\Facades\Utils;
-use App\Helpers\Messages;
-use Silver\Core\Bootstrap\Facades\Request;
-use Silver\Core\Env;
+use App\Facades\Messages;
 use Silver\Database\Model;
 use Silver\Http\Curl;
 use Silver\Http\Redirect;
 use Silver\Http\Session;
 
 /**
- * Class RestCallHelper
+ * Class RequestApiHelper
  * @author Nicolás Marulanda P.
  */
-class RestCallHelper {
+class RequestApiHelper extends ApiHelper {
     
-    private static $HTTP_STATUS_OK                    = 200;
+    private $httpRequestStatus;
     
-    private static $HTTP_STATUS_NO_CONTENT            = 204;
+    private $messageError;
     
-    private static $HTTP_STATUS_BAD_REQUEST           = 400;
-    
-    private static $HTTP_STATUS_INTERNAL_SERVER_ERROR = 500;
-    
-    private        $httpRequestStatus;
-    
-    private        $messageError;
-    
-    /**
-     * RestCallHelper constructor.
-     */
     public function __construct() {
         $this->httpRequestStatus = 0;
         $this->messageError      = [];
@@ -84,6 +69,7 @@ class RestCallHelper {
                 throw new \RuntimeException('Tipo de petición no valido');
                 break;
         }
+        
         $response = $this->objectToArray($response);
         
         $this->httpRequestStatus = $this->getValueByKey($response, 'http_status', 0);
@@ -107,10 +93,6 @@ class RestCallHelper {
     
     private function createUrl($serviceName, $serviceMethod) {
         return sprintf('%1$s/api/%2$s/%3$s', URL, $serviceName, $serviceMethod);
-    }
-    
-    public function getToken() {
-        return Session::get('token', Request::input('token', ''));
     }
     
     private function formatDataToSendRequestGet($dataToSend) {
@@ -137,28 +119,6 @@ class RestCallHelper {
         ];
     }
     
-    private function objectToArray($object, $recursive = TRUE) {
-        if (is_object($object)) {
-            $object = (array)$object;
-        }
-        
-        if ($recursive && is_array($object)) {
-            $object = array_map(function($value) {
-                return $this->objectToArray($value);
-            }, $object);
-        }
-        
-        return $object;
-    }
-    
-    private function getValueByKey($response, $key, $default) {
-        if (is_array($response) && array_key_exists($key, $response)) {
-            return $response[$key];
-        }
-        
-        return $default;
-    }
-    
     public function setToken($token) {
         Session::set('token', $token);
     }
@@ -176,87 +136,6 @@ class RestCallHelper {
         }
         
         return $this->makeRequest('GET', $dataToSend, $serviceName, $serviceMethod, $returnType);
-    }
-    
-    public function makeResponse($callback) {
-        header('Content-Type: application/json');
-        $response = [];
-        $request  = $this->getDataRequest();//recibe un array
-        
-        try {
-            if (!is_callable($callback)) {
-                throw new \Exception('El método no se puede ejecutar.');
-            }
-            
-            if (!Utils::isGetRequest() && empty($request)) {
-                $response = $this->createResponseFormat(self::$HTTP_STATUS_BAD_REQUEST, 'Faltan datos en la petición.');
-            } else {
-                $dataToSend = $callback($request);
-                
-                if ($dataToSend === NULL) {
-                    $response = $this->createResponseFormat(self::$HTTP_STATUS_NO_CONTENT);
-                } else {
-                    $response = $this->createResponseFormat(self::$HTTP_STATUS_OK, $dataToSend);
-                }
-            }
-        } catch (\Exception $ex) {
-            $response = $this->createResponseFormat(self::$HTTP_STATUS_INTERNAL_SERVER_ERROR, $ex->getMessage());
-        }
-        
-        return $response;
-    }
-    
-    public function getDataRequest() {
-        return Request::input('payload');
-    }
-    
-    public function createResponseFormat($httpStatus, $dataToSend = FALSE) {
-        $payload = [
-                'debug'         => Env::get('api_debug', FALSE),
-                'version'       => Env::get('api_version', '0.0.0'),
-                'http_status'   => $httpStatus,
-                'request_limit' => Env::get('api_request_limit', 25),
-                'token'         => TokenFacade::getToken(),
-        ];
-        
-        switch ($httpStatus) {
-            case self::$HTTP_STATUS_OK:
-                $payload['payload'] = $this->formatDataToSendResponse($dataToSend);
-                break;
-            case self::$HTTP_STATUS_NO_CONTENT:
-                break;
-            default:
-                unset($payload['token']);
-                unset($payload['request_limit']);
-                $payload['errors'] = [
-                        'message' => $dataToSend,
-                        'check'   => TRUE,
-                ];
-                break;
-        }
-        
-        return $payload;
-    }
-    
-    private function formatDataToSendResponse($dataToSend) {
-        if ($dataToSend instanceof Model) {
-            $hidden = $dataToSend->getHidden();
-            $data   = $dataToSend->data();
-            
-            $dataToSend = array_filter($data, function($key) use ($hidden) {
-                return !in_array($key, $hidden);
-            }, ARRAY_FILTER_USE_KEY);
-        } elseif (is_array($dataToSend)) {
-            foreach ($dataToSend as &$value) {
-                $value = $this->formatDataToSendResponse($value);
-            }
-        }
-        
-        return $dataToSend;
-    }
-    
-    private function headerToken() {
-        header('Authorization:' . $this->getToken());
     }
     
 }
